Add a newline automatically at the end of the file
require-final-newline is a variable defined in ‘files.el’.
Its value is t
Original value was nil
Local in buffer COMMIT_EDITMSG; global value is visit-save

  This variable is safe as a file local variable if its value
  satisfies the predicate ‘symbolp’.

Documentation:
Whether to add a newline automatically at the end of the file.

A value of t means do this only when the file is about to be saved.
A value of ‘visit’ means do this right after the file is visited.
A value of ‘visit-save’ means do it at both of those times.
Any other non-nil value means ask user whether to add a newline, when saving.
A value of nil means don’t add newlines.

Certain major modes set this locally to the value obtained
from ‘mode-require-final-newline’.